SocketAddress moved to type; base structs created
